// src/types/book.ts
export interface book {
  id: number;
  title: string;
  author: string;
  category: string;
  publicationYear: string;
  description: string;
  price: number;
  quantity: number;
  images: any;
  user: any;
}
export interface bookDetail {
  title: string;
  author: string;
  category: string;
  publicationYear: string;
  description: string;
  price: number;
  quantity: number;
  images: any;
}